+++ xenon-common/src/main/java/com/vmware/xenon/common/serialization/KryoSerializers.java
@@ -74,7 +74,7 @@ public final class KryoSerializers {
     private static ThreadLocal<Kryo> kryoForDocumentPerThreadCustom;
 
     public static final long THREAD_LOCAL_BUFFER_LIMIT_BYTES = 1024 * 1024;
-    private static final int OUTPUT_BUFFER_SIZE_BYTES = 4096;
+    private static final int DEFAULT_BUFFER_SIZE_BYTES = 4096;
     private static final BufferThreadLocal bufferPerThread = new BufferThreadLocal();
 
     public static Kryo create(boolean isObjectSerializer) {
@@ -213,7 +213,7 @@ public final class KryoSerializers {
      */
     public static Output serializeAsDocument(Object o, int maxSize) {
         Kryo k = getKryoThreadLocalForDocuments();
-        Output out = new Output(getBuffer(OUTPUT_BUFFER_SIZE_BYTES), maxSize);
+        Output out = new Output(getBuffer(DEFAULT_BUFFER_SIZE_BYTES), maxSize);
         k.writeClassAndObject(out, o);
         return out;
     }
@@ -224,8 +224,8 @@ public final class KryoSerializers {
      */
     public static Output serializeDocumentForIndexing(Object o, int maxSize) {
         Kryo k = getKryoThreadLocalForDocuments();
-        byte[] buffer = getBuffer(OUTPUT_BUFFER_SIZE_BYTES);
-        Output out = new OutputWithRoot(buffer, maxSize, o);
+        byte[] buffer = getBuffer(DEFAULT_BUFFER_SIZE_BYTES);
+        OutputWithRoot out = new OutputWithRoot(buffer, maxSize, o);
         k.writeClassAndObject(out, o);
         return out;
     }
@@ -237,7 +237,7 @@ public final class KryoSerializers {
      */
     public static int serializeObject(Object o, byte[] buffer, int position) {
         Kryo k = getKryoThreadLocalForObjects();
-        Output out = new Output(buffer);
+        Output out = new Output(buffer, buffer.length);
         out.setPosition(position);
         k.writeClassAndObject(out, o);
         return out.position();
@@ -248,7 +248,7 @@ public final class KryoSerializers {
      */
     public static ByteBuffer serializeObject(Object o, int maxSize) {
         Kryo k = getKryoThreadLocalForObjects();
-        Output out = new Output(OUTPUT_BUFFER_SIZE_BYTES, maxSize);
+        Output out = new Output(DEFAULT_BUFFER_SIZE_BYTES, maxSize);
         k.writeClassAndObject(out, o);
         return ByteBuffer.wrap(out.getBuffer(), 0, out.position());
     }